+++ src/doomstat.c
@@ -159,6 +159,8 @@ int fake_contrast;
 int wipe_speed_percentage;
 // [Cherry] General ------------------------------------------------------
 int damage_shake;
+int explosion_shake;
+int shake_percentage;
 int motion_blur;
 // [Cherry] End ----------------------------------------------------------
 int s_clipping_dist_x2;